pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
]
